requirements.txt
django==5.1.7
pillow==11.1.0
django-braces==1.17.0
django-embed-video==1.4.10
pymemcache==4.0.0
django-debug-toolbar==5.1.0
redis==5.2.1
django-redisboard==8.4.0
djangorestframework==3.16.0